Remember original deep link after authentication.
After a successfull authentication, the user is not always redirected to the index page.

If the user tried to open a page different than the index page, this page is opened after authentication.
If the user just clicked the SignIn link directly, the index page is opened as a default.

As discussed here: https://github.com/adrianhurt/play-silhouette-credentials-seed/issues/6